+++ CHANGELOG.md
@@ -12,6 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - OAuth2 `authenticate` returns `tokenData`
+- OAuth2 supports confidentials applications.
+
+```ts
+const sdk = createClient({
+  host: 'https://api.dev.fibricheck.com',
+  clientId: '',
+  clientSecret: '',
+});
+```
 
 ## [v4.2.1]